docs: Basic implementation of Astro/Starlight for docs (#60)
## Proposed changes

POC for using Astro/Starlight for the docs site. After instantiating it,
I'm on the fence about Astro. There is a bunch of powerful stuff here
that we get out of the box, but there will be some stuff to learn.

That said, I think it is definitely less maintenance burden than a
full-fledged NextJS app or something.

https://aa2cf8e2.docs-2h4.pages.dev/