feat: `execute` for sub-workflows. (#86)
## Proposed changes

Add support, an example, and tests for calling sub-workflows via
`execute`, that ensures context is passed through using
`AsyncLocalStorage`. This ensures that workflows can run concurrently
without interfering with each other.

One issue is that `AsyncLocalStorage` is not available in the browser,
and I was not able to find a polyfill that I have confidence in. So for
that situation, we just use a single global context, and warn about the
fact that concurrent workflows might have strange behavior.

Here is a run of the example:

```

> @gensx-examples/reflection@0.0.0 run /Users/jeremy/code/cortexclick/gensx/examples/reflection
> NODE_OPTIONS='--enable-source-maps' tsx ./index.tsx

(node:9668) [DEP0040] DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.
(Use `node --trace-deprecation ...` to show where the warning was created)
result:
 We are a technology company using AI solutions to deliver quality products to our customers. Our development process ensures we keep up with new ideas.

Our systems use cloud structures and current frameworks to create combined solutions that support digital change. By considering different perspectives, we help stakeholders with applications that improve return on investment.

Through our approach to innovation, we create substantial solutions that produce results. Our technology combined with our customer-focused approach allows us to develop and refine quickly in this market.
```