docs: add docs skeleton and first few pages (#94)
Adds a skeleton structure for the docs (shown below). However, only the
first three pages are fleshed out and the rest of the nav is commented
out. Including the skeleton files and nav so that we have a starting
point to iterate on.

I had a more robust quickstart in mind, but I want to make some changes
to the create-gensx starter template so that it supports streaming by
default before I expand the quickstart.

<img width="1912" alt="image"
src="https://github.com/user-attachments/assets/3f9300e1-bd54-47f3-ac30-5a56544a2261"
/>