docs: adding docs and example on providers (#154)
Recreating #148 from a clean branch


Adds the concept doc for providers

Also adds a provider example that shows how to create a provider for
Firecrawl (that is discussed in the doc)

It's worth noting that the top of this doc is the same as the text from
the basic concepts page but it extends the info. Eventually we will also
want an "Available Providers" section but when written currently, it
only contains the OpenAIProvider.