feat: Add checkpoint support for StreamComponent (#155)
Adds support for checkpointing within `StreamComponent`. Special care is
taken to fork the stream so that we don't block. When the user reads the
stream, we follow along, accumulate the value, and then send the final
`completeNode` call with the complete value at the end.

Fixes https://github.com/cortexclick/gensx/issues/142 fixes
https://github.com/cortexclick/gensx/issues/136