fix: Fix checkpoint race condition by handling out of order arrival of nodes (#170)
The full graph is executed in parallel. This means that `addNode` calls
don't necessarily occur in order, though usually they do. Luckily the
context of the `parentId` is maintained correctly all the way through
the execution tree.

The result is nodes that occassionally get dropped from the execution
graph.

@danfhernandez was occasionally seeing instances where the first
component was dropped from the checkpoint. I observed this once myself
as well:

<img width="1482" alt="image"
src="https://github.com/user-attachments/assets/be189cce-289a-4d49-bb66-463f7503280a"
/>

This change adds support for handling out of order insertion of nodes
into the graph. If we try to insert a node with a parentId into the
graph but the parent hasn't arrived yet, then we put it into an
`orphaned` nodes list, and attempt to process it on subsequent `addNode`
calls when other nodes arrive. In addition, we never write out a
checkpoint when the tree is in an invalid state.

I added tests that manually insert nodes into checkpoint manager to
simulate multiple scenarios for out of order arrival to validate and
lock in this behavior.