fix: various fixes for handling of functions and streams during resolution and checkpointing (#232)
Sorry for the vague title but this change does three semi-related
things:

1. Makes sure that any `asyncIterable` is preserved during resolution.
This is done in addition to the standard `Streamable` guard. This is
necessary for cases where a component returns something like a raw
OpenAI streaming chat completion result.
2. Mask functions and native functions from the checkpoint. This was
previously failing whenever we tried to deep clone props/outputs that
contained functions.
3. Ensure that deep resolution only executes framework functions, and
not functions that might be appended to an object returned by the user.

I added tests covering all three changes. 

(3) is the meatiest and largest portion of this change. A demonstrative
example:

```tsx
const ComponentWithFunctions = gsx.Component("Component", () => {
  return { 
    someValue: "foo", // resolve should just return "foo"
    someComponent: <OtherComponent/>, // resolve should call the OtherComponent and return the result
    someFunction: () => "bar" // we don't want resolve to call someFunction!!! it should just be returned directly to the user
  }
}
```

The key here is the third entry in the object `someFunction`. This is a
user-defined function, and not a framework function. Prior to this
change deep resolution would do the following:

1. Traverse the return object recursively.
2. Execute the component and return the result
3. Execute the user-defined function and return the result ("bar" in
this case)

This isn't what we want! Users could return objects with any number of
functions. Some of those functions could be async, or have side-effects,
and it would be very very unexpected that we call those functions
(without meaningful arguments) on their behalf -- the behavior here is
effectively undefined.

The fix here is to mark any framework-defined function (Component funcs,
Provider funcs, context methods like `with*` and `run`, etc) and mark
them with a special boolean property `__gsxFramework`. Then
`deepResolve` is updated to not blindly call all functions, just
functions where `value.__gsxFramework === true`.

There is really no way around this. We have to distinguish between
user-defied functions (an SDK that a user returns will often have a
bunch of helper functions hung onto the object) that we don't want to
execute, and framework functions that __must__ be executed.

The downside here is that we will see very strange bugs if we add more
functions to the framework and forget to mark them as `__gsxFramework`.
We have no choice but to bite this bullet. Wherever possible, I made an
effort to centralize this logic. All of the `context.ts` methods
(`withNode`, etc) make sure to mark the functions they receive as
framework functions so that we don't have to worry about marking
functions at every single call site.