docs: fixing gsxChatCompletion example (#248)
Fixing the gsxChatCompletion example 

This was broken by my change in #242 but for some reason that PR got a
green build.

I realized that the test_examples.yml build pipeline wasn't working
correctly and was running only the code already in main. To improve
this, i made the following updates:
- Makes examples run whenever a package is changed to check for breaking
changes
- Changes from `pull_request` to `pull_request_target` to make it easier
to debug and fix issues.
- Adds a clean step 

Note this PR will always fail because it's running both the old and new
versions of the test_examples file.