chore: Check if deployment works. (#315)
## Proposed changes

We are trying to track down a bug where assets return a 401 immediately
after deploying with Vercel, and a redeploy resolves the problem.

I just turned off [Skew
Protection](https://vercel.com/docs/deployments/skew-protection) to see
if this might somehow be the cause, and then made some invisible changes
to force a deploy of home and docs.