feat: @gensx/mcp (#358)
## Proposed changes

Add a new package, `@gensx/mcp`, that enables the creation of MCP
contexts with a provider and a `useContext` helper. This provider
manages the lifecycle of the server (spinning it up and shutting it
down).

One thing that was required to make this work that I'm not super stoked
about: Needed to add an `onComplete` to the `Provider` interface, so
that we can perform tasks _after_ the children of a provider have full
resolved. This is necessary for us to close the connection (aka
shutdown) the MCP server once that part of the workflow has finished.

---------

Co-authored-by: Dan Hernandez <33551757+danfhernandez@users.noreply.github.com>