feat: support structured outputs via tools in the OpenAIProvider (#440)
fixes #348 

This PR updates the OpenAIProvider to make it possible to use structured
outputs with other OpenAI Compatible APIs. If the base url is from
openai or azure openai, it will use `response_format` to create the
structured output, otherwise it will use tools.

This PR also adds a new prop `structuredOutputStrategy` to allow users
to control the method used for structured outputs. By default, the param
is set to `default` but users can also specify `tools` or
`response_format`.