fix: Fix schema extraction for structured types (#481)
* Fix cli binary build and add test

Fix resolution of named Props and structured output when bundling a
workflow and producing the schema. There was a bug that was returning `{
"type": "object" }` for any output type that was not a `string` or
`Streamable`, and there was an issue resolving named types for props.
This resolved both of those problems.