feat: add the @gensx/storage package (#541)
Adds the @gensx/storage package which includes a provider and hook that
provides runtime access to blob storage. This has two modes filesystem
for local development and cloud for deployed programs that talk to a set
of REST APIs we provide over blob storage managed on behalf of the user.

Also adds a chat-memory example.

---------

Co-authored-by: evanboyle <io.evanboyle@gmail.com>