feat: Add helpers for event stream and state. (#764)
## Proposed changes

* change `emitProgress` to `publishData`
* add `publishEvent` to publish arbitrary events with a given label.
* add `publishObject` to publish arbitrary objects with a given label.
* add `createEventStream` as a helper for creating a strongly typed
event emitter
* add `createObjectStream` as a helper for creating a strongly typed
state emitter

Things to call out:
* I opted not to use JSON patch or some sort of updating mechanism for
transmitting state changes. That would make the stream a stateful thing
(must process the whole thing when reconnecting), and I don't want to
necessarily make that required yet.
* `emitProgress` becomes a bit of a second class citizen with its new
name, and I think that is the right place for it. It is fully-flexible,
but not really the happy path that we will document front and center.
* Because we are not doing anything smart about updating objects
partially, Objects and Events are really the same, just with different
names. How we process those things (the React hooks) is where they will
differ.