feat: enable checkpoint replay (#783)
WIP branch for enabling checkpoint replay. A checkpoint JSON can be
passed into a workflow as a `workflowOpt` and be used as a "cache" to
skip over components that have already executed.

Updated checkpoint construction to insert "cached" nodes into the new
checkpoint if they are found during replay.

Updated component ID generation to be deterministic to support this.
Added tests for it all. Some TODOs remaining:

1. ID generation needs more improvement. I expect we'll discover bugs
when we start writing tests against things like JS builtins (`new
Date()`), and we will probably need to have stronger and more
deterministic props hashing vs just using `JSON.stringify`.
3. I will add some tools like `gensx.createCheckpoint()` that can be
used for testing and debugging (write a checkpoint out to a file so that
it is easier to debug, etc).

---------

Co-authored-by: Jeremy Moseley <jeremy@gensx.com>