fix: Fix create-gensx test for mismatch local vs npm versions. (#854)
## Proposed changes

We were not using local versions for all gensx packages when testing
create-gensx, so since we added vercel-ai, we have been testing with
some local versions and some versions from npm. Most of the time this is
fine, but when we do breaking changes (aka _all the time_ right now)
this
[failure](https://github.com/gensx-inc/gensx/actions/runs/16307067981/job/46055306788?pr=846)
pops up.